requirements.txt
streamlit
pymupdf
pillow
pandas
openpyxl